some BL work and UI
fixed some of the ToStrings.
added functionalaty to some of the BL functions.
I think addOrder mostly works. I made sure addOrder checks the days are not occupied in the hostingUnit Diary. Right now after adding a order i change the relavent days to occupied but we should really do that in updateOrder when the guest actually accepts the offer